+++ apps/test/fixtures/default-export-props-fn.tsx
@@ -1,0 +1,16 @@
+import { Body, Html } from 'jsx-email';
+
+export const TemplateName = 'default-export-props-fn';
+
+const Template = ({ test }: { test: string }) => (
+  <Html>
+    <Body>{test}</Body>
+  </Html>
+);
+
+Template.PreviewProps = () => {
+  return { test: 'batman' };
+};
+
+// eslint-disable-next-line import/no-default-export
+export default Template;