+++ apps/test/fixtures/empty.tsx
@@ -1,0 +1,1 @@
+// This file intentionally left black. https://github.com/shellscape/jsx-email/issues/125