+++ packages/jsx-email/src/cli/commands/build.ts
@@ -121,6 +121,7 @@ const compile = async (files: string[], outDir: string) => {
   await esbuild.build({
     bundle: true,
     entryPoints: files,
+    jsx: 'automatic',
     logLevel: 'error',
     outdir: outDir,
     platform: 'node',