+++ packages/jsx-email/src/cli/preview-app/src/components/index.ts
@@ -1,7 +1,0 @@
-export * from './code';
-export * from './heading';
-export * from './icons';
-export * from './logo';
-export * from './sidebar';
-export * from './text';
-export * from './topbar';